requirements.txt
faiss-cpu>=1.7.4
torch>=2.0.0
transformers>=4.31.0
sentence-transformers>=2.2.2
scikit-learn>=1.2.2
nltk>=3.8.1
tqdm>=4.66.1
python-docx>=1.1.0
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.8
gradio>=4.0.0
